--- src/main/java/caffeine/nest_dev/common/config/CorsConfig.java
@@ -1,19 +1,0 @@
-package caffeine.nest_dev.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") // 모든 경로에 대해 CORS 허용
-                .allowedOrigins("http://localhost:3000") // 프론트엔드 애플리케이션의 URL 명시
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS") // 허용할 HTTP 메소드
-                .allowedHeaders("*") // 모든 헤더 허용
-                .allowCredentials(true) // 자격 증명(쿠키, HTTP 인증 등) 허용
-                .maxAge(3600); // Pre-flight 요청 결과를 캐시할 시간 (초 단위)
-    }
-}
